set gemset.nix file permissions to 0644
Set the permissions on gemset.nix to match those of the other files
(e.g. Gemfile.lock). Allowing read access to users other than the file
owner is important if bundix is being used with the intent of copying
the resulting derivation to a root-owned overlay as part of a nixos
configuration. Otherwise, if a non-root user tries to invoke a nix-shell
with access to the resulting derivation, they'll fail due to a lack of
read permissions.